Move Announcer class into  package.
