+++ components/api/src/main/java/com/hotels/styx/api/extension/Announcer.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api;
+package com.hotels.styx.api.extension;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
