+++ components/api/src/test/java/com/hotels/styx/api/extension/AnnouncerTest.java
@@ -13,7 +13,7 @@
   See the License for the specific language governing permissions and
   limitations under the License.
  */
-package com.hotels.styx.api;
+package com.hotels.styx.api.extension;
 
 import org.testng.annotations.Test;
